+++ src/action_state.rs
@@ -21,6 +21,10 @@ use serde::{Deserialize, Serialize};
 pub struct ActionData {
     /// Is the action pressed or released?
     pub state: ButtonState,
+    /// The `state` of the action in the `Main` schedule
+    pub update_state: ButtonState,
+    /// The `state` of the action in the `FixedMain` schedule
+    pub fixed_update_state: ButtonState,
     /// The "value" of the binding that triggered the action.
     ///
     /// See [`ActionState::value`] for more details.
@@ -136,6 +140,30 @@ impl<A: Actionlike> Default for ActionState<A> {
 }
 
 impl<A: Actionlike> ActionState<A> {
+    /// We are about to enter the `Main` schedule, so we:
+    /// - save all the changes applied to `state` into the `fixed_update_state`
+    /// - switch to loading the `update_state`
+    pub(crate) fn swap_to_update_state(&mut self) {
+        for (_action, action_datum) in self.action_data.iter_mut() {
+            // save the changes applied to `state` into `fixed_update_state`
+            action_datum.fixed_update_state = action_datum.state;
+            // switch to loading the `update_state` into `state`
+            action_datum.state = action_datum.update_state;
+        }
+    }
+
+    /// We are about to enter the `FixedMain` schedule, so we:
+    /// - save all the changes applied to `state` into the `update_state`
+    /// - switch to loading the `fixed_update_state`
+    pub(crate) fn swap_to_fixed_update_state(&mut self) {
+        for (_action, action_datum) in self.action_data.iter_mut() {
+            // save the changes applied to `state` into `update_state`
+            action_datum.update_state = action_datum.state;
+            // switch to loading the `fixed_update_state` into `state`
+            action_datum.state = action_datum.fixed_update_state;
+        }
+    }
+
     /// Updates the [`ActionState`] based on a vector of [`ActionData`], ordered by [`Actionlike::id`](Actionlike).
     ///
     /// The `action_data` is typically constructed from [`InputMap::which_pressed`](crate::input_map::InputMap),
@@ -743,7 +771,7 @@ mod tests {
         input_map.insert(Action::Run, KeyCode::KeyR);
 
         // Starting state
-        let input_streams = InputStreams::from_world(&app.world, None);
+        let input_streams = InputStreams::from_world(app.world(), None);
         action_state.update(input_map.process_actions(&input_streams, ClashStrategy::PressAll));
 
         assert!(!action_state.pressed(&Action::Run));
@@ -755,7 +783,7 @@ mod tests {
         app.press_input(KeyCode::KeyR);
         // Process the input events into Input<KeyCode> data
         app.update();
-        let input_streams = InputStreams::from_world(&app.world, None);
+        let input_streams = InputStreams::from_world(app.world(), None);
 
         action_state.update(input_map.process_actions(&input_streams, ClashStrategy::PressAll));
 
@@ -776,7 +804,7 @@ mod tests {
         // Releasing
         app.release_input(KeyCode::KeyR);
         app.update();
-        let input_streams = InputStreams::from_world(&app.world, None);
+        let input_streams = InputStreams::from_world(app.world(), None);
 
         action_state.update(input_map.process_actions(&input_streams, ClashStrategy::PressAll));
 
@@ -818,7 +846,7 @@ mod tests {
         let mut action_state = ActionState::<Action>::default();
 
         // Starting state
-        let input_streams = InputStreams::from_world(&app.world, None);
+        let input_streams = InputStreams::from_world(app.world(), None);
         action_state
             .update(input_map.process_actions(&input_streams, ClashStrategy::PrioritizeLongest));
         assert!(action_state.released(&Action::One));
@@ -828,7 +856,7 @@ mod tests {
         // Pressing One
         app.press_input(Digit1);
         app.update();
-        let input_streams = InputStreams::from_world(&app.world, None);
+        let input_streams = InputStreams::from_world(app.world(), None);
 
         action_state
             .update(input_map.process_actions(&input_streams, ClashStrategy::PrioritizeLongest));
@@ -849,7 +877,7 @@ mod tests {
         // Pressing Two
         app.press_input(Digit2);
         app.update();
-        let input_streams = InputStreams::from_world(&app.world, None);
+        let input_streams = InputStreams::from_world(app.world(), None);
 
         action_state
             .update(input_map.process_actions(&input_streams, ClashStrategy::PrioritizeLongest));