+++ src/clashing_inputs.rs
@@ -491,7 +491,7 @@ mod tests {
             app.press_input(Digit3);
             app.update();
 
-            let input_streams = InputStreams::from_world(&app.world, None);
+            let input_streams = InputStreams::from_world(app.world(), None);
 
             assert_eq!(
                 resolve_clash(
@@ -523,7 +523,7 @@ mod tests {
 
             input_map.handle_clashes(
                 &mut action_data,
-                &InputStreams::from_world(&app.world, None),
+                &InputStreams::from_world(app.world(), None),
                 ClashStrategy::PrioritizeLongest,
             );
 
@@ -552,7 +552,7 @@ mod tests {
 
             input_map.handle_clashes(
                 &mut action_data,
-                &InputStreams::from_world(&app.world, None),
+                &InputStreams::from_world(app.world(), None),
                 ClashStrategy::PrioritizeLongest,
             );
 
@@ -574,7 +574,7 @@ mod tests {
             app.update();
 
             let action_data = input_map.process_actions(
-                &InputStreams::from_world(&app.world, None),
+                &InputStreams::from_world(app.world(), None),
                 ClashStrategy::PrioritizeLongest,
             );