+++ tests/mouse_motion.rs
@@ -47,20 +47,20 @@ fn raw_mouse_move_events() {
     let mut app = test_app();
     app.insert_resource(InputMap::new([(AxislikeTestAction::X, MouseMoveAxis::Y)]));
 
-    let mut events = app.world.resource_mut::<Events<MouseMotion>>();
+    let mut events = app.world_mut().resource_mut::<Events<MouseMotion>>();
     events.send(MouseMotion {
         delta: Vec2::new(0.0, 1.0),
     });
 
     app.update();
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
     assert!(action_state.pressed(&AxislikeTestAction::X));
 }
 
 #[test]
 fn mouse_move_discrete_mocking() {
     let mut app = test_app();
-    let mut events = app.world.resource_mut::<Events<MouseMotion>>();
+    let mut events = app.world_mut().resource_mut::<Events<MouseMotion>>();
     assert_eq!(events.drain().count(), 0);
 
     app.press_input(MouseMoveDirection::UP);
@@ -72,26 +72,26 @@ fn mouse_move_discrete_mocking() {
 #[test]
 fn mouse_move_single_axis_mocking() {
     let mut app = test_app();
-    let mut events = app.world.resource_mut::<Events<MouseMotion>>();
+    let mut events = app.world_mut().resource_mut::<Events<MouseMotion>>();
     assert_eq!(events.drain().count(), 0);
 
     let input = MouseMoveAxis::X;
     app.send_axis_values(input, [-1.0]);
 
-    let mut events = app.world.resource_mut::<Events<MouseMotion>>();
+    let mut events = app.world_mut().resource_mut::<Events<MouseMotion>>();
     assert_eq!(events.drain().count(), 1);
 }
 
 #[test]
 fn mouse_move_dual_axis_mocking() {
     let mut app = test_app();
-    let mut events = app.world.resource_mut::<Events<MouseMotion>>();
+    let mut events = app.world_mut().resource_mut::<Events<MouseMotion>>();
     assert_eq!(events.drain().count(), 0);
 
     let input = MouseMove::default();
     app.send_axis_values(input, [1.0, 0.0]);
 
-    let mut events = app.world.resource_mut::<Events<MouseMotion>>();
+    let mut events = app.world_mut().resource_mut::<Events<MouseMotion>>();
     // Dual axis events are split out
     assert_eq!(events.drain().count(), 2);
 }
@@ -107,7 +107,7 @@ fn mouse_move_buttonlike() {
     ]));
 
     for action in ButtonlikeTestAction::variants() {
-        let input_map = app.world.resource::<InputMap<ButtonlikeTestAction>>();
+        let input_map = app.world().resource::<InputMap<ButtonlikeTestAction>>();
         // Get the first associated input
         let input = input_map.get(action).unwrap().first().unwrap().clone();
         let direction = Reflect::as_any(input.as_ref())
@@ -117,7 +117,7 @@ fn mouse_move_buttonlike() {
         app.press_input(*direction);
         app.update();
 
-        let action_state = app.world.resource::<ActionState<ButtonlikeTestAction>>();
+        let action_state = app.world().resource::<ActionState<ButtonlikeTestAction>>();
         assert!(action_state.pressed(action), "failed for {input:?}");
     }
 }
@@ -138,7 +138,7 @@ fn mouse_move_buttonlike_cancels() {
     // Correctly flushes the world
     app.update();
 
-    let action_state = app.world.resource::<ActionState<ButtonlikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<ButtonlikeTestAction>>();
 
     assert!(!action_state.pressed(&ButtonlikeTestAction::Up));
     assert!(!action_state.pressed(&ButtonlikeTestAction::Down));
@@ -156,42 +156,42 @@ fn mouse_move_single_axis() {
     let input = MouseMoveAxis::X;
     app.send_axis_values(input, [1.0]);
     app.update();
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
     assert!(action_state.pressed(&AxislikeTestAction::X));
 
     // -X
     let input = MouseMoveAxis::X;
     app.send_axis_values(input, [-1.0]);
     app.update();
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
     assert!(action_state.pressed(&AxislikeTestAction::X));
 
     // +Y
     let input = MouseMoveAxis::Y;
     app.send_axis_values(input, [-1.0]);
     app.update();
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
     assert!(action_state.pressed(&AxislikeTestAction::Y));
 
     // -Y
     let input = MouseMoveAxis::Y;
     app.send_axis_values(input, [-1.0]);
     app.update();
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
     assert!(action_state.pressed(&AxislikeTestAction::Y));
 
     // 0
     let input = MouseMoveAxis::Y;
     app.send_axis_values(input, [0.0]);
     app.update();
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
     assert!(!action_state.pressed(&AxislikeTestAction::Y));
 
     // No value
     let input = MouseMoveAxis::Y;
     app.send_axis_values(input, []);
     app.update();
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
     assert!(!action_state.pressed(&AxislikeTestAction::Y));
 }
 
@@ -207,7 +207,7 @@ fn mouse_move_dual_axis() {
     app.send_axis_values(input, [5.0, 0.0]);
     app.update();
 
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
 
     assert!(action_state.pressed(&AxislikeTestAction::XY));
     assert_eq!(action_state.value(&AxislikeTestAction::XY), 5.0);
@@ -229,7 +229,7 @@ fn mouse_move_discrete() {
     app.send_axis_values(input, [0.0, -2.0]);
     app.update();
 
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
 
     assert!(action_state.pressed(&AxislikeTestAction::XY));
     // This should be a unit length, because we're working with a VirtualDPad
@@ -259,7 +259,7 @@ fn mouse_drag() {
     app.press_input(MouseButton::Right);
     app.update();
 
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
 
     assert!(action_state.pressed(&AxislikeTestAction::XY));
     assert_eq!(