+++ tests/mouse_wheel.rs
@@ -45,7 +45,7 @@ fn raw_mouse_scroll_events() {
         MouseScrollDirection::UP,
     )]));
 
-    let mut events = app.world.resource_mut::<Events<MouseWheel>>();
+    let mut events = app.world_mut().resource_mut::<Events<MouseWheel>>();
     events.send(MouseWheel {
         unit: MouseScrollUnit::Pixel,
         x: 0.0,
@@ -54,14 +54,14 @@ fn raw_mouse_scroll_events() {
     });
 
     app.update();
-    let action_state = app.world.resource::<ActionState<ButtonlikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<ButtonlikeTestAction>>();
     assert!(action_state.pressed(&ButtonlikeTestAction::Up));
 }
 
 #[test]
 fn mouse_scroll_discrete_mocking() {
     let mut app = test_app();
-    let mut events = app.world.resource_mut::<Events<MouseWheel>>();
+    let mut events = app.world_mut().resource_mut::<Events<MouseWheel>>();
     assert_eq!(events.drain().count(), 0);
 
     app.press_input(MouseScrollDirection::UP);
@@ -73,26 +73,26 @@ fn mouse_scroll_discrete_mocking() {
 #[test]
 fn mouse_scroll_single_axis_mocking() {
     let mut app = test_app();
-    let mut events = app.world.resource_mut::<Events<MouseWheel>>();
+    let mut events = app.world_mut().resource_mut::<Events<MouseWheel>>();
     assert_eq!(events.drain().count(), 0);
 
     let input = MouseScrollAxis::X;
     app.send_axis_values(input, [-1.0]);
 
-    let mut events = app.world.resource_mut::<Events<MouseWheel>>();
+    let mut events = app.world_mut().resource_mut::<Events<MouseWheel>>();
     assert_eq!(events.drain().count(), 1);
 }
 
 #[test]
 fn mouse_scroll_dual_axis_mocking() {
     let mut app = test_app();
-    let mut events = app.world.resource_mut::<Events<MouseWheel>>();
+    let mut events = app.world_mut().resource_mut::<Events<MouseWheel>>();
     assert_eq!(events.drain().count(), 0);
 
     let input = MouseScroll::default();
     app.send_axis_values(input, [-1.0]);
 
-    let mut events = app.world.resource_mut::<Events<MouseWheel>>();
+    let mut events = app.world_mut().resource_mut::<Events<MouseWheel>>();
     // Dual axis events are split out
     assert_eq!(events.drain().count(), 2);
 }
@@ -108,7 +108,7 @@ fn mouse_scroll_buttonlike() {
     ]));
 
     for action in ButtonlikeTestAction::variants() {
-        let input_map = app.world.resource::<InputMap<ButtonlikeTestAction>>();
+        let input_map = app.world().resource::<InputMap<ButtonlikeTestAction>>();
         // Get the first associated input
         let input = input_map.get(action).unwrap().first().unwrap().clone();
         let direction = Reflect::as_any(input.as_ref())
@@ -118,7 +118,7 @@ fn mouse_scroll_buttonlike() {
         app.press_input(*direction);
         app.update();
 
-        let action_state = app.world.resource::<ActionState<ButtonlikeTestAction>>();
+        let action_state = app.world().resource::<ActionState<ButtonlikeTestAction>>();
         assert!(action_state.pressed(action), "failed for {input:?}");
     }
 }
@@ -139,7 +139,7 @@ fn mouse_scroll_buttonlike_cancels() {
     // Correctly flushes the world
     app.update();
 
-    let action_state = app.world.resource::<ActionState<ButtonlikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<ButtonlikeTestAction>>();
 
     assert!(!action_state.pressed(&ButtonlikeTestAction::Up));
     assert!(!action_state.pressed(&ButtonlikeTestAction::Down));
@@ -157,42 +157,42 @@ fn mouse_scroll_single_axis() {
     let input = MouseScrollAxis::X;
     app.send_axis_values(input, [1.0]);
     app.update();
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
     assert!(action_state.pressed(&AxislikeTestAction::X));
 
     // -X
     let input = MouseScrollAxis::X;
     app.send_axis_values(input, [-1.0]);
     app.update();
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
     assert!(action_state.pressed(&AxislikeTestAction::X));
 
     // +Y
     let input = MouseScrollAxis::Y;
     app.send_axis_values(input, [1.0]);
     app.update();
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
     assert!(action_state.pressed(&AxislikeTestAction::Y));
 
     // -Y
     let input = MouseScrollAxis::Y;
     app.send_axis_values(input, [-1.0]);
     app.update();
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
     assert!(action_state.pressed(&AxislikeTestAction::Y));
 
     // 0
     let input = MouseScrollAxis::Y;
     app.send_axis_values(input, [0.0]);
     app.update();
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
     assert!(!action_state.pressed(&AxislikeTestAction::Y));
 
     // No value
     let input = MouseScrollAxis::Y;
     app.send_axis_values(input, []);
     app.update();
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
     assert!(!action_state.pressed(&AxislikeTestAction::Y));
 }
 
@@ -208,7 +208,7 @@ fn mouse_scroll_dual_axis() {
     app.send_axis_values(input, [5.0, 0.0]);
     app.update();
 
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
 
     assert!(action_state.pressed(&AxislikeTestAction::XY));
     assert_eq!(action_state.value(&AxislikeTestAction::XY), 5.0);
@@ -230,7 +230,7 @@ fn mouse_scroll_discrete() {
     app.send_axis_values(input, [0.0, -2.0]);
     app.update();
 
-    let action_state = app.world.resource::<ActionState<AxislikeTestAction>>();
+    let action_state = app.world().resource::<ActionState<AxislikeTestAction>>();
 
     assert!(action_state.pressed(&AxislikeTestAction::XY));
     // This should be a unit length, because we're working with a VirtualDPad